# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x402 Bazaar API",
    "description": "The decentralized API marketplace where AI agents pay with USDC via HTTP 402 protocol. Free endpoints return data directly. Paid endpoints return HTTP 402 with payment details (wallet address, amount in USDC, supported chains). To use paid endpoints, send USDC on Base to the specified wallet, then retry with X-Payment-TxHash header. 69+ total endpoints available — this schema covers the 30 most popular ones.",
    "version": "2.0.0",
    "contact": {
      "name": "x402 Bazaar",
      "url": "https://x402bazaar.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/Wintyx57/x402-backend/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://x402-api.onrender.com",
      "description": "Production (Base mainnet)"
    }
  ],
  "paths": {
...
    "/api/weather": {
      "get": {
        "operationId": "getWeather",
        "summary": "Weather data for any city (0.02 USDC)",
        "description": "Get current weather data (temperature, wind speed, conditions) for any city worldwide. Costs 0.02 USDC.",
        "parameters": [
          { "name": "city", "in": "query", "required": true, "description": "City name (e.g., Paris, Tokyo, New York)", "schema": { "type": "string", "maxLength": 100 } }
        ],
        "responses": {
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "200": { "description": "Weather data", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
    "/api/crypto": {
      "get": {
        "operationId": "getCryptoPrice",
        "summary": "Cryptocurrency price data (0.02 USDC)",
        "description": "Get current price, market cap, and 24h change for any cryptocurrency. Costs 0.02 USDC.",
        "parameters": [
          { "name": "coin", "in": "query", "required": true, "description": "Cryptocurrency ID (e.g., bitcoin, ethereum, solana)", "schema": { "type": "string" } }
        ],
        "responses": {
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "200": { "description": "Crypto price data", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
    "/api/joke": {
      "get": {
        "operationId": "getRandomJoke",
        "summary": "Random joke (0.01 USDC)",
        "description": "Returns a random joke (setup + punchline or single-liner). Costs 0.01 USDC.",
        "responses": {
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "200": { "description": "A random joke", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
    "/api/search": {
      "get": {
        "operationId": "webSearch",
        "summary": "Web search for LLMs (0.005 USDC)",
        "description": "Returns clean, structured web search results optimized for AI consumption. Costs 0.005 USDC.",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "description": "Search query", "schema": { "type": "string", "maxLength": 200 } }
        ],
        "responses": {
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "200": { "description": "Search results", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
    "/api/scrape": {
      "get": {
        "operationId": "scrapeUrl",
        "summary": "Universal URL scraper (0.005 USDC)",
        "description": "Scrapes any URL and returns clean Markdown content, optimized for LLM consumption. Costs 0.005 USDC.",
        "parameters": [
          { "name": "url", "in": "query", "required": true, "description": "URL to scrape", "schema": { "type": "string", "format": "uri" } }
        ],
        "responses": {
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "200": { "description": "Scraped content in Markdown", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
    "/api/twitter": {
      "get": {
        "operationId": "getTwitterData",
        "summary": "Twitter/X data (0.005 USDC)",
        "description": "Get Twitter/X user profiles, individual tweets, or search results. Provide one of: user, tweet, or search. Costs 0.005 USDC.",
        "parameters": [
          { "name": "user", "in": "query", "description": "Twitter username (without @)", "schema": { "type": "string" } },
          { "name": "tweet", "in": "query", "description": "Tweet ID or URL", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "description": "Search keyword for tweets", "schema": { "type": "string" } }
        ],
        "responses": {
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "200": { "description": "Twitter data", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
    "/api/image": {
      "get": {
        "operationId": "generateImage",
        "summary": "AI image generation DALL-E 3 (0.05 USDC)",
        "description": "Generate an image from a text prompt using DALL-E 3. Returns a URL. Costs 0.05 USDC.",
        "parameters": [
          { "name": "prompt", "in": "query", "required": true, "description": "Image generation prompt", "schema": { "type": "string", "maxLength": 1000 } },
          { "name": "size", "in": "query", "description": "Image size", "schema": { "type": "string", "enum": ["1024x1024", "1024x1792", "1792x1024"], "default": "1024x1024" } }
        ],
        "responses": {
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "200": { "description": "Generated image URL", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PaymentRequired": {
        "type": "object",
        "description": "Returned when payment is required. Contains details for making a USDC payment on Base blockchain.",
        "properties": {
          "error": { "type": "string", "example": "Payment Required" },
          "x402": {
            "type": "object",
            "properties": {
              "version": { "type": "string", "example": "1" },
              "payload": {
                "type": "object",
                "properties": {
                  "amount": { "type": "string", "description": "Amount in USDC micro-units (6 decimals)", "example": "5000" },
                  "amountHuman": { "type": "string", "description": "Human-readable USDC amount", "example": "0.005 USDC" },
                  "recipient": { "type": "string", "description": "Wallet address to send USDC to", "example": "0xfb1c478BD5567BdcD39782E0D6D23418bFda2430" },
                  "chain": { "type": "string", "example": "base" },
                  "chainId": { "type": "integer", "example": 8453 },
                  "token": { "type": "string", "example": "USDC" },
                  "tokenAddress": { "type": "string", "example": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913" }
                }
              }
            }
          },
          "instructions": { "type": "string", "description": "How to complete payment and retry the request" }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true }
        },
        "description": "Response varies by endpoint. Always includes success:true and endpoint-specific data."
      }
    },
    "responses": {
      "PaymentRequired": {
        "description": "Payment required. Send USDC on Base to the specified wallet address, then retry with X-Payment-TxHash header.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/PaymentRequired" }
          }
        }
      }
    }
  }
}