# routes/services.js
// routes/services.js — GET /services, GET /search, GET /api/services, GET /api/services/:id,
//                     GET /api/activity, GET /api/services/activity, GET /api/health-check,
//                     DELETE /api/admin/services/:id, POST /api/admin/services/:id/verify

const express = require('express');
const logger = require('../lib/logger');
const { fetchWithTimeout } = require('../lib/payment');
const { ServiceSearchSchema, ServiceListQuerySchema } = require('../schemas/index');
const { verifyService } = require('../lib/service-verifier');
const { safeUrl } = require('../lib/safe-url');

function createServicesRouter(supabase, logActivity, paymentMiddleware, paidEndpointLimiter, dashboardApiLimiter, adminAuth) {
    const router = express.Router();

    // --- LISTE DES SERVICES (0.05 USDC) ---
    router.get('/services', paidEndpointLimiter, paymentMiddleware(50000, 0.05, "List Services"), async (req, res) => {
        const { data, error } = await supabase
            .from('services')
            .select('*')
            .order('created_at', { ascending: true });

        if (error) {
            logger.error('Supabase', '/services error:', error.message);
            return res.status(500).json({ error: 'Failed to fetch services' });
        }

        res.json({
            success: true,
            count: data.length,
            data
        });
    });

    // --- RECHERCHE DE SERVICES (0.05 USDC) ---
    router.get('/search', paidEndpointLimiter, paymentMiddleware(50000, 0.05, "Search Services"), async (req, res) => {
        // Validate query parameters using Zod
        const parseResult = ServiceSearchSchema.safeParse({ q: req.query.q || '' });

        if (!parseResult.success) {
            const errors = parseResult.error.errors.map(err => err.message).join(', ');
            return res.status(400).json({ error: errors });
        }

        const query = parseResult.data.q;

        // Sanitize: escape special Postgres LIKE characters
        const sanitized = query.replace(/[%_\\]/g, '\\$&');

        // Recherche floue sur name et description
        const pgSafe = sanitized.replace(/[(),."']/g, '');
        const { data, error } = await supabase
            .from('services')
            .select('*')
            .or(`name.ilike.%${pgSafe}%,description.ilike.%${pgSafe}%`);

        if (error) {
            logger.error('Supabase', '/search error:', error.message);
            return res.status(500).json({ error: 'Search failed' });
        }

        logActivity('search', `Search "${query}" -> ${data.length} result(s)`);

        res.json({
            success: true,
            query,
            count: data.length,
            data
        });
    });

    // --- API services (gratuit, pour le dashboard) ---
    router.get('/api/services', dashboardApiLimiter, async (req, res) => {
        const { data, error } = await supabase
            .from('services')
            .select('*')
            .order('created_at', { ascending: false });

        if (error) {
            logger.error('Supabase', '/api/services error:', error.message);
            return res.status(500).json({ error: 'Failed to fetch services' });
        }
        res.json(data);
    });

    // --- API activity log (gratuit, pour le dashboard) ---
    router.get('/api/activity', dashboardApiLimiter, async (req, res) => {
        const { data, error } = await supabase
            .from('activity')
            .select('type, detail, amount, created_at')
            .order('created_at', { ascending: false })
            .limit(50);

        if (error) {
            logger.error('Supabase', '/api/activity error:', error.message);
            return res.status(500).json({ error: 'Failed to fetch activity' });
        }

        // Mapper pour compatibilite dashboard (time)
        const activity = (data || []).map(a => ({
            type: a.type,
            detail: a.detail,
            amount: Number(a.amount),
            time: a.created_at,
        }));

        res.json(activity);
    });

    // --- SERVICES ACTIVITY (Last call timestamps) ---
    router.get('/api/services/activity', dashboardApiLimiter, async (req, res) => {
        try {
            const { data, error } = await supabase
                .from('activity')
                .select('detail, created_at')
                .eq('type', 'api_call')
                .order('created_at', { ascending: false })
                .limit(200);

            if (error) {
                logger.error('Supabase', '/api/services/activity error:', error.message);
                return res.status(500).json({ error: 'Failed to fetch activity' });
            }

            // Map detail patterns to endpoints and find latest timestamp
            const activityMap = {};
            const endpointPatterns = [
                { pattern: /Web Search API/i, endpoint: '/api/search' },
                { pattern: /Scraper API/i, endpoint: '/api/scrape' },
                { pattern: /Twitter API/i, endpoint: '/api/twitter' },
                { pattern: /Weather API/i, endpoint: '/api/weather' },
                { pattern: /Crypto (?:Price )?API/i, endpoint: '/api/crypto' },
                { pattern: /(?:Random )?Joke API/i, endpoint: '/api/joke' },
                { pattern: /Image (?:Generation )?API/i, endpoint: '/api/image' },
            ];

            for (const row of (data || [])) {
                for (const { pattern, endpoint } of endpointPatterns) {
                    if (pattern.test(row.detail) && !activityMap[endpoint]) {
                        activityMap[endpoint] = row.created_at;
                    }
                }
            }

            res.json(activityMap);
        } catch (err) {
            logger.error('Activity', err.message);
            res.status(500).json({ error: 'Internal error' });
        }
    });

    // --- GET SINGLE SERVICE BY ID ---
    router.get('/api/services/:id', dashboardApiLimiter, async (req, res) => {
        try {
            const { id } = req.params;

            // Validate UUID format
            const uuidRegex = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
            if (!uuidRegex.test(id)) {
                return res.status(400).json({ error: 'Invalid service ID format' });
            }

            const { data, error } = await supabase
                .from('services')
                .select('*')
                .eq('id', id)
                .single();

            if (error || !data) {
                return res.status(404).json({ error: 'Service not found' });
            }

            res.json(data);
        } catch (err) {
            logger.error('Supabase', `/api/services/:id error: ${err.message}`);
            res.status(500).json({ error: 'Internal server error' });
        }
    });

    // --- HEALTH CHECK (service URLs) ---
    const healthCache = new Map();
    const HEALTH_TTL = 10 * 60 * 1000; // 10 minutes

    // Cleanup expired healthCache entries every 30 min
    setInterval(() => {
        const now = Date.now();
        for (const [key, val] of healthCache) {
            if (now - val.timestamp > HEALTH_TTL * 3) healthCache.delete(key);
        }
    }, 30 * 60 * 1000).unref();

    router.get('/api/health-check', dashboardApiLimiter, async (req, res) => {
        try {
            const { data: services, error } = await supabase
                .from('services')
                .select('url')
                .order('created_at', { ascending: false });

            if (error) {
                return res.status(500).json({ error: 'Failed to fetch services' });
            }

            // Deduplicate base URLs
            const urls = [...new Set((services || []).map(s => s.url).filter(Boolean))];

            const results = {};
            const toCheck = [];

            // Check cache first
            for (const url of urls) {
                const cached = healthCache.get(url);
                if (cached && (Date.now() - cached.timestamp < HEALTH_TTL)) {
                    results[url] = cached.status;
                } else {
                    toCheck.push(url);
                }
            }

            // Batch check remaining URLs (batches of 10)
            for (let i = 0; i < toCheck.length; i += 10) {
                const batch = toCheck.slice(i, i + 10);
                const checks = batch.map(async (url) => {
                    try {
                        await safeUrl(url);
                    } catch {
                        results[url] = 'blocked';
                        return;
                    }
                    try {
                        const controller = new AbortController();
                        const timeout = setTimeout(() => controller.abort(), 5000);
                        const response = await fetch(url, {
                            method: 'HEAD',
                            signal: controller.signal,
                            redirect: 'follow',
                        });
                        clearTimeout(timeout);
                        // Status 402 is normal for x402 (payment required = online)
                        const status = (response.status >= 200 && response.status < 500) ? 'online' : 'offline';
                        healthCache.set(url, { status, timestamp: Date.now() });
                        results[url] = status;
                    } catch {
                        healthCache.set(url, { status: 'offline', timestamp: Date.now() });
                        results[url] = 'offline';
                    }
                });
                await Promise.all(checks);
            }

            res.json(results);
        } catch (err) {
            logger.error('Health Check', err.message);
            res.status(500).json({ error: 'Health check failed' });
        }
    });

    // --- ADMIN: RE-VERIFY SERVICE ---
    if (adminAuth) {
        router.post('/api/admin/services/:id/verify', adminAuth, async (req, res) => {
            const { id } = req.params;

            // Fetch service from DB
            const { data: services, error: fetchErr } = await supabase
                .from('services')
                .select('*')
                .eq('id', id)
                .limit(1);

            if (fetchErr || !services || services.length === 0) {
                return res.status(404).json({ error: 'Service not found' });
            }

            const service = services[0];
            const report = await verifyService(service.url);

            // Update verified_status in DB
            await supabase
                .from('services')
                .update({ verified_status: report.verdict, verified_at: new Date().toISOString() })
                .eq('id', id);

            logActivity('admin', `Re-verified "${service.name}": ${report.verdict}`);
            logger.info('Admin', `Re-verified "${service.name}" (${id.slice(0, 8)}): ${report.verdict}`);

            res.json({ success: true, service: service.name, report });
        });
    }

    // --- ADMIN: DELETE SERVICE BY ID ---
    if (adminAuth) {
        router.delete('/api/admin/services/:id', adminAuth, async (req, res) => {
            const { id } = req.params;
            const { data, error } = await supabase
                .from('services')
                .delete()
                .eq('id', id)
                .select();

            if (error) {
                logger.error('Admin', `DELETE /api/admin/services/${id} error: ${error.message}`);
                return res.status(500).json({ error: 'Failed to delete service' });
            }

            if (!data || data.length === 0) {
                return res.status(404).json({ error: 'Service not found' });
            }

            logActivity('admin', `Deleted service: ${data[0].name} (${id})`);
            logger.info('Admin', `Deleted service: ${data[0].name} (${id})`);
            res.json({ success: true, deleted: data[0] });
        });

        // Debug: step-by-step payment verification (admin only, disabled in production)
        router.get('/api/admin/debug-verify', adminAuth, async (req, res) => {
            if (process.env.NODE_ENV === 'production') return res.status(404).json({ error: 'Not found' });
            const txHash = req.query.tx;
            const chainKey = req.query.chain || 'base';
            if (!txHash) return res.status(400).json({ error: 'Missing ?tx= parameter' });

            const { getChainConfig } = require('../lib/chains');
            const chain = getChainConfig(chainKey);
            const steps = [];
            try {
                const normalizedTxHash = txHash.toLowerCase().trim();
                steps.push({ step: 'normalize', txHash: normalizedTxHash, length: normalizedTxHash.length });

                const serverAddress = process.env.WALLET_ADDRESS?.toLowerCase();
                steps.push({ step: 'serverAddress', address: serverAddress, env_set: !!process.env.WALLET_ADDRESS });

                steps.push({ step: 'chain', chainKey, label: chain.label, rpcUrls: chain.rpcUrls, usdcContract: chain.usdcContract });

                // Fetch receipt
                const rpcUrl = chain.rpcUrls?.[0] || chain.rpcUrl;
                const receiptRes = await fetch(rpcUrl, {
                    method: 'POST',
                    headers: { 'Content-Type': 'application/json' },
                    body: JSON.stringify({ jsonrpc: '2.0', method: 'eth_getTransactionReceipt', params: [normalizedTxHash], id: 1 }),
                });
                const receiptData = await receiptRes.json();
                const receipt = receiptData.result;
                steps.push({ step: 'receipt', found: !!receipt, status: receipt?.status, blockNumber: receipt?.blockNumber, logsCount: receipt?.logs?.length });

                if (!receipt) { steps.push({ step: 'FAIL', reason: 'No receipt found' }); return res.json({ steps }); }

                // Block confirmations
                const blockRes = await fetch(rpcUrl, {
                    method: 'POST',
                    headers: { 'Content-Type': 'application/json' },
                    body: JSON.stringify({ jsonrpc: '2.0', method: 'eth_blockNumber', params: [], id: 2 }),
                });
                const { result: currentBlockHex } = await blockRes.json();
                const currentBlock = parseInt(currentBlockHex, 16);
                const txBlock = parseInt(receipt.blockNumber, 16);
                steps.push({ step: 'confirmations', currentBlock, txBlock, confirmations: currentBlock - txBlock });

                // Parse logs
                const TRANSFER_TOPIC = '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef';
                for (const log of (receipt.logs || [])) {
                    if (log.topics[0] === TRANSFER_TOPIC && log.topics.length >= 3) {
                        const logAddr = log.address.toLowerCase();
                        const usdcMatch = logAddr === chain.usdcContract.toLowerCase();
                        const toAddr = '0x' + log.topics[2].slice(26).toLowerCase();
                        const walletMatch = toAddr === serverAddress;
                        const amount = BigInt(log.data);
                        steps.push({
                            step: 'transfer_log',
                            logAddress: logAddr,
                            usdcMatch,
                            to: toAddr,
                            walletMatch,
                            amount: Number(amount),
                            amountUsdc: Number(amount) / 1e6,
                        });
                    }
                }
                steps.push({ step: 'DONE' });
            } catch (err) {
                steps.push({ step: 'ERROR', message: err.message });
            }
            res.json({ steps });
        });

        // Diagnostic: which community-agent env vars are set (no values exposed)
        router.get('/api/admin/env-check', adminAuth, (req, res) => {
            const keys = [
                'AGENT_PRIVATE_KEY', 'TELEGRAM_BOT_TOKEN', 'TELEGRAM_CHAT_ID', 'TELEGRAM_CHANNEL_ID',
                'GEMINI_API_KEY', 'DISCORD_WEBHOOK_URL', 'OPENAI_API_KEY', 'ENABLE_COMMUNITY_AGENT',
                'DEVTO_API_KEY', 'MAX_BUDGET_USDC',
            ];
            const result = {};
            for (const k of keys) {
                result[k] = !!process.env[k];
            }
            res.json(result);
        });
    }

    return router;
}

module.exports = createServicesRouter;